dist_requirements.txt
sixgill-clients==0.2.4
swimlane
requests==2.25.1